legacy/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-customer-support"
version = "0.1.0"
description = "Multi-Agent system for Technical Customer Support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.29.0",
    "box>=0.1.5",
    "google-adk>=0.0.1",
    "litellm>=1.0.0",
    "duckduckgo-search>=5.0.0",
    "openai>=1.69.0",
    "pandas>=2.2.3",
    "pgvector>=0.4.0",
    "psycopg2-binary>=2.9.10",
    "pyaml>=25.1.0",
    "pypdf>=5.4.0",
    "python-box>=7.3.2",
    "sqlalchemy>=2.0.40",
    "redis>=6.0.0"
]

[tool.setuptools]
packages = ["src"]

